+++ packages/gov-ui-kit/src/core/utils/invariant/invariant.ts
@@ -1,0 +1,12 @@
+export const invariantError = 'Invariant';
+
+export function invariant(condition: boolean, message: string): asserts condition {
+    if (!condition) {
+        const error = new Error(message);
+        error.name = invariantError;
+
+        throw error;
+    }
+
+    return;
+}