+++ packages/gov-ui-kit/src/modules/components/proposal/index.ts
@@ -1,2 +1,3 @@
 export * from './proposalActions';
 export * from './proposalDataListItem';
+export * from './proposalVoting';