+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingBreakdownMultisig/index.ts
@@ -1,0 +1,4 @@
+export {
+    ProposalVotingBreakdownMultisig,
+    type IProposalVotingBreakdownMultisigProps,
+} from './proposalVotingBreakdownMultisig';