+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingBreakdownToken/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingBreakdownToken, type IProposalVotingBreakdownTokenProps } from './proposalVotingBreakdownToken';