+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingContainer/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingContainer, type IProposalVotingContainerProps } from './proposalVotingContainer';