+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingDefinitions.ts
@@ -1,0 +1,13 @@
+export enum ProposalVotingTab {
+    BREAKDOWN = 'BREAKDOWN',
+    VOTES = 'VOTES',
+    DETAILS = 'DETAILS',
+}
+
+export enum ProposalVotingStatus {
+    ACTIVE = 'ACTIVE',
+    PENDING = 'PENDING',
+    ACCEPTED = 'ACCEPTED',
+    REJECTED = 'REJECTED',
+    UNREACHED = 'UNREACHED',
+}