+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingDetails/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingDetails, type IProposalVotingDetailsProps } from './proposalVotingDetails';