+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingStage/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingStage, type IProposalVotingStageProps } from './proposalVotingStage';