+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingStageStatus/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingStageStatus, type IProposalVotingStageStatusProps } from './proposalVotingStageStatus';