+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingTabs/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingTabs, type IProposalVotingTabsProps } from './proposalVotingTabs';