+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingVotes/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingVotes, type IProposalVotingVotesProps } from './proposalVotingVotes';