+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingVotes/proposalVotingVotes.tsx
@@ -1,0 +1,8 @@
+import { type ITabsContentProps, Tabs } from '../../../../../core';
+import { ProposalVotingTab } from '../proposalVotingDefinitions';
+
+export interface IProposalVotingVotesProps extends Omit<ITabsContentProps, 'value'> {}
+
+export const ProposalVotingVotes: React.FC<IProposalVotingVotesProps> = (props) => {
+    return <Tabs.Content value={ProposalVotingTab.VOTES} {...props} />;
+};